Implemented controllers and updated services

+++ src/main/java/com/github/prajjwal/florio/dto/ServiceBookingResponseDto.java
@@ -24,7 +24,7 @@ public class ServiceBookingResponseDto {
     private Double finalPrice;
     private Instant createdAt;
     private Instant completedAt;
-    private Double rating;
+    private Integer rating;
     private String feedback;
 
 }
